Move intercepted-request to new folder alongside the used structures
