+++ traffic-sniffer/index.ts
@@ -3,4 +3,4 @@ export * from "./lib/request-metadata/sniffer/sniffer-controller";
 export * from "./lib/request-metadata/sniffer-manager/sniffer-manager";
 export * from "./lib/request-metadata/sniffer-manager/sniffer-manager-controller";
 export * from "./lib/request-metadata/intercepted-request";
-export * from "./lib/request-metadata/intercepted-requests";
+export * from "./lib/request-metadata/intercepted-request/intercepted-requests";
